// Fase 3/Evidencias Grupales/Evidencias Del sistema/src/models/cliente.model.ts
export interface Cliente {
  id?: string;
  telefono: string;        
  nombreReal: string;       
  nombreAlias: string;     
  fechaNacimiento: string;  
  generoSeleccionado: string; 
  descripcion: string;     
  intereses: string[];      
  cualidades: string[];   
  fotosAdicionales?: string[];
  mostrarNombreReal?: boolean;
}